+++ stories/amo/PromotedBadge.js
@@ -2,47 +2,49 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 
-import IconPromotedBadge, {
-  IconPromotedBadgeBase,
-} from 'ui/components/IconPromotedBadge';
+import PromotedBadge, { PromotedBadgeBase } from 'amo/components/PromotedBadge';
+import type { Props as PromotedBadgeProps } from 'amo/components/PromotedBadge';
 
 import Provider from '../setup/Provider';
 
-const render = ({ category = 'line', size = 'large' } = {}) => {
-  return <IconPromotedBadge category={category} size={size} />;
+const render = (moreProps: $Shape<PromotedBadgeProps> = {}) => {
+  const props = {
+    category: 'line',
+    size: 'large',
+    ...moreProps,
+  };
+  return <PromotedBadge {...props} />;
 };
 
-storiesOf('IconPromotedBadge', module)
-  .addParameters({
-    component: IconPromotedBadgeBase,
-  })
+storiesOf('PromotedBadge', module)
+  .addParameters({ component: PromotedBadgeBase })
   .addDecorator((story) => <Provider story={story()} />)
   .addWithChapters('all variants', {
     chapters: [
       {
         sections: [
           {
-            title: 'Line, size="large"',
+            title: 'category="line", size="large"',
             sectionFn: () => render({ size: 'large' }),
           },
           {
-            title: 'Line, size="small"',
+            title: 'category="line", size="small"',
             sectionFn: () => render({ size: 'small' }),
           },
           {
-            title: 'Recommended, size="large"',
+            title: 'category="recommended", size="large"',
             sectionFn: () => render({ category: 'recommended', size: 'large' }),
           },
           {
-            title: 'Recommended, size="small"',
+            title: 'category="recommended", size="small"',
             sectionFn: () => render({ category: 'recommended', size: 'small' }),
           },
           {
-            title: 'Verified, size="large"',
+            title: 'category="verified", size="large"',
             sectionFn: () => render({ category: 'verified', size: 'large' }),
           },
           {
-            title: 'Verified, size="small"',
+            title: 'category="verified", size="small"',
             sectionFn: () => render({ category: 'verified', size: 'small' }),
           },
         ],
